third_party/blink/web_tests/external/wpt/css/mediaqueries/forced-colors-expected.txt

This is a testharness.js-based test.
PASS Should be parseable in a CSS stylesheet: '(forced-colors)'
FAIL Should be parseable in a CSS stylesheet: '(forced-colors: none)' assert_true: expected true got false
FAIL Should be parseable in a CSS stylesheet: '(forced-colors: active)' assert_true: expected true got false
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: 0)'
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: no-preference)'
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: 10px)'
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: active 0)'
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: none active)'
PASS Should not be parseable in a CSS stylesheet: '(forced-colors: active/none)'
PASS Should be parseable in JS: '(forced-colors)'
FAIL Should be parseable in JS: '(forced-colors: none)' assert_true: expected true got false
FAIL Should be parseable in JS: '(forced-colors: active)' assert_true: expected true got false
PASS Should not be parseable in JS: '(forced-colors: 0)'
PASS Should not be parseable in JS: '(forced-colors: no-preference)'
PASS Should not be parseable in JS: '(forced-colors: 10px)'
PASS Should not be parseable in JS: '(forced-colors: active 0)'
PASS Should not be parseable in JS: '(forced-colors: none active)'
PASS Should not be parseable in JS: '(forced-colors: active/none)'
FAIL Check that none evaluates to false in the boolean context assert_equals: expected true but got false
Harness: the test ran to completion.
